frontend-utils: Add new package and moved custom toml parsing utilities there
